# GH-98831: Remove all remaining DISPATCH() calls from bytecodes.c