--- a/Python/bytecodes.c
+++ b/Python/bytecodes.c
@@ -930,10 +930,10 @@ dummy_func(
                 PUSH(value);
             }
             else {
-                    PyObject *exc_type = Py_NewRef(Py_TYPE(exc_value));
-                    PyObject *exc_traceback = PyException_GetTraceback(exc_value);
-                    _PyErr_Restore(tstate, exc_type, Py_NewRef(exc_value), exc_traceback);
-                    goto exception_unwind;
+                PyObject *exc_type = Py_NewRef(Py_TYPE(exc_value));
+                PyObject *exc_traceback = PyException_GetTraceback(exc_value);
+                _PyErr_Restore(tstate, exc_type, Py_NewRef(exc_value), exc_traceback);
+                goto exception_unwind;
             }
         }
 
@@ -2429,10 +2429,12 @@ dummy_func(
                     STACK_SHRINK(1);
                     Py_DECREF(cond);
                 }
-                else if (err == 0)
+                else if (err == 0) {
                     JUMPBY(oparg);
-                else
+                }
+                else {
                     goto error;
+                }
             }
         }
 
@@ -2456,8 +2458,9 @@ dummy_func(
                     STACK_SHRINK(1);
                     Py_DECREF(cond);
                 }
-                else
+                else {
                     goto error;
+                }
             }
         }
 
@@ -2639,7 +2642,7 @@ dummy_func(
             STACK_SHRINK(1);
             Py_DECREF(it);
             JUMPBY(INLINE_CACHE_ENTRIES_FOR_ITER + oparg + 1);
-         end_for_iter_list:
+        end_for_iter_list:
         }
 
         // stack effect: ( -- __0)