--- a/Python/generated_cases.c.h
+++ b/Python/generated_cases.c.h
@@ -930,10 +930,10 @@
                 PUSH(value);
             }
             else {
-                    PyObject *exc_type = Py_NewRef(Py_TYPE(exc_value));
-                    PyObject *exc_traceback = PyException_GetTraceback(exc_value);
-                    _PyErr_Restore(tstate, exc_type, Py_NewRef(exc_value), exc_traceback);
-                    goto exception_unwind;
+                PyObject *exc_type = Py_NewRef(Py_TYPE(exc_value));
+                PyObject *exc_traceback = PyException_GetTraceback(exc_value);
+                _PyErr_Restore(tstate, exc_type, Py_NewRef(exc_value), exc_traceback);
+                goto exception_unwind;
             }
             DISPATCH();
         }
@@ -2443,10 +2443,12 @@
                     STACK_SHRINK(1);
                     Py_DECREF(cond);
                 }
-                else if (err == 0)
+                else if (err == 0) {
                     JUMPBY(oparg);
-                else
+                }
+                else {
                     goto error;
+                }
             }
             DISPATCH();
         }
@@ -2470,8 +2472,9 @@
                     STACK_SHRINK(1);
                     Py_DECREF(cond);
                 }
-                else
+                else {
                     goto error;
+                }
             }
             DISPATCH();
         }
@@ -2654,7 +2657,7 @@
             STACK_SHRINK(1);
             Py_DECREF(it);
             JUMPBY(INLINE_CACHE_ENTRIES_FOR_ITER + oparg + 1);
-         end_for_iter_list:
+        end_for_iter_list:
             DISPATCH();
         }