@gvanrossum 2022-11-10
@brandtbucher When you have the benchmark results can you just post them here? So Mark won't have to ask for them.